requirements.txt
Flask>=2.0
boto3>=1.26
requests>=2.28
PyYAML>=6.0
gunicorn>=20.1
